Remove the e from HttpDecodeErrors
HttpDecodeError String doesn't have an Exception instance, because there
is no Exception instance for String. Therefore, getResponseBodyUnsafe
didn't work with httpJson, which was basically its entire purpose.

To fix this, we make HttpDecodeError hold a String internally and ask
callers to marshal it themselves by passing ByteString-> (Either
(NonEmpty String) a to httpDecode.

An alternative considered was to use a newtype as necessary in each
decoding case. For example, an AesonDecodeError wrapper over String with
an Exception instance that just displays it. Similar newtypes would
effectively be the suggested way to use httpDecode with functions that
use non-Exception types in Left.

This alternative was discarded because it is more machinery for no
benefit. Callers always have to get their decoding error types to a
String for us somehow. Previously, it was via an Exception instance;
now, it'd be via the decoding function itself. In the previous approach,
Exception types were handled for free, and non-Exception types would
thus require a newtype and instance.

In the current approach, String types are handled for free, and other
types require an extra function call (and displayException works great).
Given that libraries seem to rarely use Exception types for their pure
decoders, and often use String, biasing this way seems like the correct
choice.