Throw HttpStatusError first in getResponseBodyUnsafe
Users of getResponseBodyUnsafe are typically just looking for a
successful value and to crash otherwise. For unsuccessful responses, the
body was unlikely to parse as they expected, but throwing whatever is
the Left error hides the fact that it was unsuccessful to begin with.

This patch will throw a new HttpStatusError instead. It still holds the
Either response body, so access to all the same information is still
present; the error should just be clearer.

This doesn't change getResponseBody, which will return whatever body is
there (Either or not) regardless of status. The assumption is that, if
you are going to check the body for Either yourself, you can/will also
check statusIsSuccessful . getResponseStatus too. For this expected
usage, I added statusIsSuccessful to our HTTP re-exports.