Change withApp to not require a Pool
Separating loadPool and withApp was an unnecessary complexity that arose
as a way to solve us wanting to truncate tables as part of app setup in
tests.

We thought we would need a handle on the pool first to do that, so
loading it once and passing it to the user's loadApp was a way to do
that.

Two things were discovered:

- Lots of times we don't truncate at all
- We can easily read the pool off a loaded App with HasSqlPool

So now we just expect loadApp to include loading its pool, if necessary.
And we offer a new hook to run something on that pool provided the app
HasSqlPool.

Users of this library that would previously do:

    main = do
      app <- loadApp =<< makePostgresPool
      -- ...

    loadApp pool = App pool
      <*> ...
      <*> ...

Can now do

    main = do
      app <- loadApp
      -- ...

    loadApp = App
      <$> makePostgresPool
      <*> ...

In tests, where one was doing:

    withAppTruncate load = withApp setup
     where
      setup pool = do
        runSqlPool truncateTables pool
        load pool

Can now do

    withAppTruncate = withAppSql truncateTables

Users of withApp would require no changes, once their loadApp itself is
updated.

There's going to be a behavior change with regard to when things happen.

Before: the Pool would be loaded once before all spec items, the
truncation (for example) and loading of App would happen before each
spec item.

After: the App (and its Pool) will be loaded once before all spec items,
and (only) the truncation would happen before each spec item.

I think this is an improvement.