Separate test suite.
The `VersionSpec` specs have been extracted from the `spec` test suite /
target into a `gittest` target.

The `gittest` target will only be built if the `git` flag is passed.
This is the **default**, but it allows us to ignore this test suite in
specific circumstances.

The reason for this split is that [we can't run
git](https://github.com/commercialhaskell/stackage/issues/6172) when
stackage is running our test suite.