Make PGSTATEMENTTIMEOUT a more specific type
In our API, we want to set a statement timeout that is very close to,
but just short of, our overall request timeout of 10s. Concretely, we've
been using 9.9 seconds (or 9900ms). With the PGSTATEMENTTIMEOUT being
written as a newtype over Int, we couldn't support that and had to
work around it.

Shipping this will allow us to use PGSTATEMENTTIMEOUT=9900ms and not
have to do anything further. This is good because the API method sets
its statement_timeout every runDB call, which adds a little overhead,
while the Freckle.App.Database method is in connection setup.