Update to hspec-unit-formatter-1.1.0.0
This version removed runJUnitSpec in favor of functions better-suited to
the common case (e.g. hspecJUnit). This, however, is not the common
case.

We are running three suites separately and evaluating everything once at
the end. Therefore, we have to dig around in hspec internals a little
more and re-implement runJUNitSpec ourselves.