Add Freckle-specific prelude
**Why?**

In removing `classy-prelude` from our monorepo, a ton of import ergonomics have
been lost. This attempts to recover some of those ergonomics and also be the
starting point for our very own prelude.

This was mostly inspired by the things I had to import a gazillion types when
removing `classy-prelude` from a few of our libraries. Before hitting the
library that has the most instances, I thought it'd be nice to have something
like this in place.

**Things to look out for**

 - Did I export too much?
 - Are there glaring omissions?
 - Did I pick the wrong thing in some case (e.g. is `Control.Monad.Reader` the
 - "right" reader?)
 - Should this move into our monorepo?
 - Should I get a dep from another package? All new packages were hidden deps,
 - which is nice, but should I have instead just squashed RIO + Prelude?

**Note on reviewing**

This is likely controversial so I'll be sure to eyes from all interested
backenders.