Simplify getRouteName, getTraceId to getTags
By accepting a generic producer of [(Text, Text)], callers can supply
the same route (and trace_id) values we are using today, but also be
able to supply any others (without blowing up the interface).

Concretely, this will be used to add a Freckle Team as the "owner" tag
to everything we emit from our Haskell APIs.