Open up ports
Using a named service container only works if your steps are also in a
container. To use steps on the VM you want to open ports to the service
container and access it at localhost.

There are ways to use dynamic ports for this setup too, but grabbing the
conventional one should be fine.

See https://github.com/actions/example-services/blob/main/.github/workflows/postgres-service.yml