Treat MEMCACHED_SERVERS="" as a disabled client
This feels just as, if not more, reasonable than accepting either of ""
or "memcached://" to mean "defaults", but it's also practically more
convenient to operate.

At least in our systems, it is easier to unconditionally supply a
variable and conditionally make it empty, than it is to conditionally
supply a variable.