Run Database pool creation in MonadLogger
This is how it is meant to work: creation of the pool captures the
logging context and will use it to log database activity. Doing this
allowed us to address some TODOs for logging token handling as well. I
also removed some TODOs that were _TODONE_ already, and some dead code.

To recover the previous behavior, end-users can import and use
`runNoLoggingT` themselves; but they are encouraged to use their real
logging context instead.