Re-implement FrontRow.App.Env on envparse
When we first created this module, `envparse` was in an unmaintained
state. It had fallen out of Stackage, had a few bugs, and there was no
response from the maintainers. Therefore, we elected to cut our own port
of it with intentions to extract and publish as a replacement.

We never extracted, and `envparse` is now maintained. We're also
starting to extract other libraries that we'd like to expose environment
variable parsers from (e.g. Blammo) that can compose together into
larger parsers. We can't do that if freckle-app holds the machinery for
parsing. So rather than extract now, we should migrate back to
`envparse` and can target that from any libraries.

For end-users, the following changes will be required (examples of all
of which are visible in this diff):

1. If you use `Parser a` or `Reader a`, make it (e.g.) `Parser Error a`
2. Instead of `Env.parse $ ...`, use `Env.parse id $ ...`
3. Instead of `var x X nonEmpty`, use `var (x <=< nonempty) X mempty`

   Not that `(str <=< nonempty)` is redundant and can just be
   `nonempty`.

4. Instead of `var (f <$> x) ...`, do `f <$> var x ...`

   This may not always be required. Also note that if you're using `def`
   here, the type of the default will need to match `x` and not `f`.

5. If you were relying on parsed variables remaining in the environment
   after parsing, add `keep` to the modifiers.

6. If you were using `handleEither`, that's been removed.

   This is not possible to implement with `envparse`. You likely need to
   move to a two-stage process where you parse an intermediate type and
   then further validate/normalize it and throw your own errors as
   appropriate.