Add Freckle.App.Env.kept
Modifies a `Parser` such that all variables act as if they were defined
using `keep`. The past-tense naming was chosen to match the similar
function `prefixed`.

By default, read variables are removed from the environment. This is
often problematic (e.g. in tests that repeatedly load an app and re-read
the environment), and the security benefits are minor. Doing this
variable by variable is error-prone and often outside of the control of
where it matters (e.g. parser defined in this library, ultimately used
in some app). This function will make it easy for the app to decide at
parse time if it wants all variables kept all the time.