Require a bracketed loadApp function
Currently freckle-app, expects a function like:

```hs
loadApp :: IO App
```

Which it uses in `Freckle.App.runApp` and `Freckle.App.Test.withApp`.
This is problematic for Apps that make use of resources that only expose
their own usage in a bracket-like way:

```hs
-- withStatsClient :: (StatsClient -> m a) -> m a

data App = App
  { ...
  , appStatsClient :: StatsClient
  }

loadApp :: IO App
loadApp = do
  withStatsClient $ \appStatsClient -> do
    -- ???
```

And even if we could work around it, we'd be losing any safety for
resource-cleanup that that usage is intending to provide.

To support this, we can expect a bracket-like function ourselves,

```hs
loadApp :: (App -> IO a) -> IO a
loadApp f = do
  withStatsClient $ \appStatsClient -> do
    f App {..}
```

No problem.

So this commit updates `runApp` and `Test.withApp` to do exactly that,
then updates the Memcache test to take advantage of it, mostly as an
example.

I chose to do this in a backwards-incompatible way rather than make
alternative functions. This method is trivial to implement if you're
using the old way (and the reverse is impossible).

```hs
loadApp :: (App -> IO a) -> IO a
loadApp f = do
  app <- loadApp'
  f app

loadApp' :: IO App
loadApp' = do
  -- your old code
```

I had to update the main documentation header, so I also cleaned up some
out of date, but unrelated aspects therein.