Export tagsL
The automatic adding of tags to the thread context doesn't work for a
web application that forks a new thread (with an empty context) for
every request.

This can be solved pretty easily by setting up the context yourself in a
WAI middleware, but it requires pulling the tags off of the app type,
and so we need a function to access them.

The existing Lens function maintains existing encapsulation, and
composes well with `statsClientL`:

```hs
statsTags :: HasStatsClient env => env -> [(Text, Text)]
statsTags = app ^.. statsClientL . tagsL . traverse
```

This can be used with a Middleware like `addThreadContext` from Blammo.