Remove AppVersion
By using Dotenv, we can set `DD_VERSION=$(git ...)` to get an up to date
SHA version in development. Production is already shipping an explicitly
`DD_VERSION` value on deployment. The only thing `AppVersion` is
providing us is the committed-at data -- which we don't use.

I'm retaining the `HasAppVersion` class for Bugsnag for now, but may
simplify that away as well -- now that we collect metadata from the
ambient `Stats` tags, which we know will have `version` as coming from
`DD_VERSION` already.