Group SqlError Timeouts by Request URL
A timeout can happen basically anywhere, so having them all grouped
together is not useful for triaging. Segregating them at least by
Request URL would be an improvement for us. Note that if the Request or
URL is missing, we fall back to grouping them all together. This means
you should ensure the `updateEventFromWaiRequest` `BeforeNotify` runs
before this one.