Drop query string from URL used to group timeouts
Query strings will vary from call to call, especially because we include
an `echo` GET param that is unique to every request. This means we'd be
effectively making unique timeout Errors every time.

Naively stripping the query string works (vs parsing the URL), since
we're not using this for anything critical, and will ensure this groups
by domain-path only. Hashes should not be involved for a Backend API,
though we could strip that too if we find cases.