Actually publish Gauge values
As written, the `withGauge` machinery was incr/decrementing a stateful
gauge but then just never sending it anywhere.

This commit adds the missing `publish` step to the calls, which
necessitated changing the `Gauges` type slightly to also store the
metric name. I also moved to the seemingly more appropriate `inc` and
`dec` `EKG` functions, vs `add 1` and `subtract 1`. It probably doesn't
matter, but this may handle edge cases better, such as decrementing a
0/negative value.