Auto-commit offsets (#115)
* Auto-commit offsets

Manually committing offsets when processing messages might imply that something particular happens if committing an offset is skipped, such as the message being sent again by the broker. However, this is not usually the case: as soon as an offset for a later message is committed, a broker will not send an earlier skipped message in the topic partition.

Skipping messages is fine for the only application currently consuming messages using Kafka. In a future hypothetical application in which every message must be seen exactly or at least once, manually committing offsets is not sufficient due to the aforementioned behavior. There are strategies to rewind the offset in order to re-consume data, but they are currently beyond the scope of the needs for this library.

* Fix lint warning

* Bump version