Create Freckle.App.Async
We make heavy use of the local thread context for maintain contextual
attributes on any log messages. It can be very easy to spawn a new
thread which lacks the parent's context. If that thread does any
logging, those details will be missing. The three places touched by this
commit all had this bug.

To fix it now, and avoid it later, `Freckle.App.Async` offers versions
of functions like `UnliftIO.Async` that do the Right Thing in this
regard. It is also a natural home for higher-level, concurrency-related
constructs, such as creating persistent processes using the `immortal`
library.