Centralize adding TraceId to the logging context
This commit adds some functions to OpenTelemetry for working with the
current trace-id as converted to Datadog's format and for the purposes
of setting the log context. It's then re-used in the Middleware from
which it was extracted.

This will allow us to do the same thing in other persistent-process
runners, such as Kafka consumers, just like we do in APIs through the
middleware.