Add script to manually upload docs
Once we started depending on libkafka, the Haddocks no longer build on
Hackage itself, so we must manually upload them. This script was used to
build and publish docs for 1.10.1.0.

We can't use `stack haddock` for this, because it doesn't produce
documentation suitable for Hackage (the links are all wrong). `cabal
haddock` has some special sauce there.

To avoid requiring a non-Stack Haskell toolchain in a Stack-based
project, we run the `cabal` portions of this script via the `haskell`
docker image, with an appropriate GHC. It would be easy to remove this,
or make it optional, in the future, though.