Add script for manually uploading Haddocks to Hackage
Hackage's automatic Haddock builds don't work for freckle-app, since we need the Kafka library on the system. This PR adds a script to build and upload them manually. Since this needs cabal haddock (stack haddock doesn't work for this use-case), I do add a cabal.project file, with documentation enabled. I found it quite easy to set up nix and use that to run it, as described in the comment.

Ideally, we'd set this up automatically in the release.yml workflow, but that would either require migrating all of CI to cabal (and/or nix) or Stack to fix its haddock so we can extend our stack-upload-action. The latter actually seems more expedient, as there has been lots of activity on my bug report.