Add tracing to Memcached operations
Wraps get and set with `cache.get` and `cache.set` spans respectively.
The get span includes the key as an attribute and the set span includes
the key, value, and expiration.

This adds a `MonadTracer` constraint to all caching functions, but that
should be manageable since most apps using `freckle-app` and caching
will have it there for `Freckle.App.Database.runDB` anyway.