Add tracing to Kafka.Consumer operations
This adds an outer span, and inner spans for decoding and handling the
message.

In order for spans to be correctly marked as error with exception
details, you need to "throw through" the `inSpan` function, so I had to
do that refactor too. This ultimately worked out simpler by avoiding
nested eithers and moving error handling concerns out of the main body.