Remove unsafe decodeUtf8 in cache tracing
We attempt to add a `value` attribute to traces in caching, and we use
`decodeUtf8` to do so. We're encountering errors decoding some values as
UTF-8. These values in particular should decode fine, however the
library shouldn't be assuming that anyway. It should support any
`ByteString` value whatsoever. Failing here also prevents values from
being `set` at all, which can have drastic performance ramifications
that far outweigh the value of this observability-related feature.

To resolve specifically that, we use `decodeUtf8With lenientDecode`
instead. As part of that, we extract a dedicated function and also
implement a length limit, as these values could be any size. This also
gives us a handy place to document the situation.