Fix db metrics
Previously, the active_pool_connections metric was incremented _before_
the connection was obtained from the pool. This means that if there was
a queue of requests awaiting connection availablity, the metric value
would increase.

The number of enqueued requests is still valuable to know though so we
pick a metric with a better name (queries_enqueued_and_processing)
and use it in the place of active_pool_connections, and we set
active_pool_connections in another way which (I believe) will accurately
capture this.