Upload Haddocks from GitHub
Since implementing Kafka functionality, Hackage has been unable to build
our Haddocks due to the system library requirement that introduced.
Since then, we've been sporadically uploading docs using the
`docs/upload` script, which relied on (nix and) cabal to use its
`--haddock-for-hackage` flag.

With the release of stack-v2.15, we can now:

- Use its new `--haddock --haddock-for-hackage` options
- Use the new `stack upload -d` option
- Use `HACKAGE_KEY` to authorize `stack` for uploads

So we can do all of this on CI, without incurring the need to use cabal
instead of Stack there. We also don't really need `stack-upload-action`
anymore, so that's been removed.

I decided to add an additional build within `release.yml`, rather than:

1. Conditionally release from `ci.yml` after building, or
2. Attempting to upload artifacts from `ci.yml` for use in `release.yml`

The reason for this is:

- It made this the simplest change to accomplish the goal
- It's somewhat valuable to add a requirement the thing builds before
  releasing, which wasn't the case before
- The addition of Haddocks makes the build 23m cached-or-not, which we
  don't want to do except during release, and is fine to do post-merge,
  which is not holding anyone up