Explicitly loop in `runConsumer`
We were relying on `immortal` to handle the looping for `runConsumer` which calls the `onFinish` after every loop. This makes our strategy of calling a synchronous offset commit happen on every loop instead of just on error/shutdown of the consumer.

This changes to explicitly loop in the consumer so that `onFinish` only gets called on shutdown/crashes.