Freckle.App.Http.Cache
Facilities to wrap a request-response function into one that caches
responses the way a CDN or browser would.

- Respect `Cache-Control` and `Vary` on the request
- Respect `Cache-Control`, `Age`, `Expires` on the response
- Only cache methods and status that support it
- Key the request by URL + `Vary`
- On a stale cache with `ETag` request with `If-None-Match` and respect
  a 304 response
- Store the response in GZipped form to reduce cache size

The concrete storage mechanism for the cache is abstract, as well as how
to de/serialize into it, but a `Memcached` implementation is provided.

NOTE: since the response bodies are fully read to be cached, this may
not work well for large or chunked responses.