Set service.name in client spans
Datadog has a lot of trace- and overview-level graphs that break things
down by the services involved. Without setting these names, the spans
that represent remote services don't appear as such and aren't broken
out in such graphs.

Using generic naming for the services ("database", "kafka") isn't ideal,
as different users of this library talking to distinct databases will
appear as talking to the same service, but it's the best we can do for
now without a large lift to introduce settings types throughout our
stacks so that clients can specify names. This is still an improvement
worth making in the meantime, we can always improve this later.