Don't be Datadog-specific in OpenTelemetry module
The existing OpenTelemetry modules here were careful to assume and
support Datadog as a backend. This was because we thought you had to in
order for it to function. That was either not the case, or is no longer
the case, as Datadog seems able to handle default id values and even
prefers to encode them as hex in its own UI. Converting ids to a
`Word64` value is no longer necessary, and actively confusing to see.

Because of this, we can simplify and remove some of our code here. This
change deprecates various things to nudge users a certain way when
upgrading.