Use matrix.include to decide if haddocks need uploading
Originally, I thought only `freckle-kafka` needed this, but since
`freckle-app` is still depending on everything, nothing about it's
system needs have changed, so it too needs haddocks manually built
(still).

Now that we have two packages like this, we can move to an
`include`-style, which is much easier to maintain.