Add `TextualEnum` `newtype` wrapper (#186)
This is an experiment that I'm putting out for review to see what the team
thinks. I won't feel bad if we don't like it.

**What is it?**

`TextualEnum` is a `newtype` wrapper meant to provide a lot of boilerplate
serialization/deserialization instances that we keep defining and redefining as
we introduce new types and services.

**How's it work?**

The core idea is that if you can give your type an `Enum` instance, a `Bounded`
instance, and a way to convert to text, then that should be enough to provide
string-type (i.e. textual) representations of it.

So `toText` tells us how to serialize the value, and we can look through the
`Bounded`/`Enum` instance values to parse it "from text".

`toText` is preferable from `fromText` because, for bounded/enums, totality is
straightforward.

**EnumName**

I've also included `EnumName` because we've started introducing codec/schema
definitions that can be named.

**Example**

You can see the specs (`PrimaryColor`) for an example.

**Benefits**

This should reduce a great deal of boilerplate, and, if used, will reduce a lot
of ad-hoc-ry that exists across our in-the-wild definitions.