fix(freckle-http): use TTL in set operation
The backend-agnostic cache logic encodes TTL into the cached data,
checks it on read, and ignores/deletes it if expired.

For a backend that supports expiry itself (e.g. memcached), this should
not be necessary. And, at times of heavy load, the extra `delete`
operations can cause performance problems.

To address this, we update the `set` member of `HttpCache` to accept the
TTL value too. Then, in the memcached implementation, we use it to set
with the key that TTL, meaning expired values will simply not be
retrieved by `get` at all, and no `delete`s will occur (modulo races).

We do still retain the backend-agnostic logic, since the state
implementation (for example) does not support expiring keys and so
expiration still needs to be checked externally for such backends.

Since the signature of `HttpCache.set` is changing, this must be a major
version bump, even though it is unlikely to disrupt any users in
practice.